+++ test/models/spredcast.js
@@ -8,18 +8,19 @@ var user2;
 var user3;
 var cast1;
 var cast2;
+var token;
 
 describe('Testing Spredcast models', function () {
   before(function (done) {
-    common.userModel.createPassword(fixture.user.email, fixture.user.password, fixture.user.pseudo, "", "", function (err, cUser) {
+    common.userModel.createPassword(fixture.user.email, fixture.user.password, fixture.user.pseudo, '', '', function (err, cUser) {
       if (err) {
         done(err);
       } else {
-        common.userModel.createPassword(fixture.user2.email, fixture.user2.password, fixture.user2.pseudo, "", "", function (err, cUser2) {
+        common.userModel.createPassword(fixture.user2.email, fixture.user2.password, fixture.user2.pseudo, '', '', function (err, cUser2) {
           if (err) {
             done(err);
           } else {
-            common.userModel.createPassword(fixture.user3.email, fixture.user3.password, fixture.user3.pseudo, "", "", function (err, cUser3) {
+            common.userModel.createPassword(fixture.user3.email, fixture.user3.password, fixture.user3.pseudo, '', '', function (err, cUser3) {
               if (err) {
                 done(err);
               } else {
@@ -93,7 +94,7 @@ describe('Testing Spredcast models', function () {
 
   describe('spredCastModel.userCanJoin()', function () {
     it('Should authorize only caster when stream is not started', function (done) {
-      common.spredCastModel.userCanJoin(cast1._id, user._id, function (err, authorisation, fCast) {
+      common.spredCastModel.userCanJoin(cast1._id, user._id, true, function (err, authorisation, fCast) {
         if (err) {
           done(err);
         } else {
@@ -103,7 +104,7 @@ describe('Testing Spredcast models', function () {
       });
     });
     it('Should authorize only caster when stream is not started', function (done) {
-      common.spredCastModel.userCanJoin(cast1._id, user2._id, function (err, authorisation, fCast) {
+      common.spredCastModel.userCanJoin(cast1._id, user2._id, false, function (err, authorisation, fCast) {
         if (err) {
           done(err);
         } else {
@@ -114,8 +115,6 @@ describe('Testing Spredcast models', function () {
     });
   });
 
-  describe('spredcastModel.updateState()', function () {
-  });
 
   describe('spredCastModel.getByUrl()', function () {
     it('Should find the researched cast', function (done) {
@@ -128,13 +127,115 @@ describe('Testing Spredcast models', function () {
         }
       });
     });
+  });
 
-    it('Should return if no cast is found', function (done) {
-      common.spredCastModel.getByUrl(' fpe', function (err, fCast) {
+  describe('spredCastModel.updateState()', function () {
+    it('Should update state of the spredCast', function (done) {
+      common.spredCastModel.updateState(cast1._id, 1, function (err) {
         if (err) {
           done(err);
         } else {
-          expect(fCast).to.be.null;
+          common.spredCastModel.getById(cast1._id, function (err, fCast) {
+            if (err) {
+              done(err);
+            } else {
+              expect(fCast.state).to.equal(1);
+              done();
+            }
+          });
+        }
+      });
+    });
+
+    it('Should update state of the spredCast', function (done) {
+      common.spredCastModel.updateState(cast2._id, 1, function (err) {
+        if (err) {
+          done(err);
+        } else {
+          common.spredCastModel.getById(cast2._id, function (err, fCast) {
+            if (err) {
+              done(err);
+            } else {
+              expect(fCast.state).to.equal(1);
+              done();
+            }
+          });
+        }
+      });
+    });
+
+    it('Sould now allow user2 to access cast1', function (done) {
+      common.spredCastModel.userCanJoin(cast1._id, user2._id, false, function (err, authorisation, fCast) {
+        if (err) {
+          done(err);
+        } else {
+          expect(authorisation).to.be.true;
+          done();
+        }
+      });
+    });
+
+    it('Sould now allow user2 to access cast2', function (done) {
+      common.spredCastModel.userCanJoin(cast2._id, user2._id, false, function (err, authorisation, fCast) {
+        if (err) {
+          done(err);
+        } else {
+          expect(authorisation).to.be.true;
+          done();
+        }
+      });
+    });
+
+    it('Sould not allow user3 to access cast12', function (done) {
+      common.spredCastModel.userCanJoin(cast2._id, user3._id, false, function (err, authorisation, fCast) {
+        if (err) {
+          done(err);
+        } else {
+          expect(authorisation).to.be.false;
+          done();
+        }
+      });
+    });
+  });
+
+  describe('spredCastModel.updateUserCount()', function () {
+    it('Should update the user count number', function (done) {
+      common.spredCastModel.updateUserCount(cast1._id, 2, function (err) {
+        if (err) {
+          done(err);
+        } else {
+          common.spredCastModel.getById(cast1._id, function (err, fCast) {
+            if (err) {
+              done(err);
+            } else {
+              expect(fCast.userCount).to.equal(2);
+              done();
+            }
+          });
+        }
+      });
+    });
+  });
+
+  describe('castTokenModel.createNew()', function () {
+    it('Should create a new cast token with an existing user', function (done) {
+      common.castTokenModel.createNew(client, user, cast1, true, function (err, cToken) {
+        if (err) {
+          done(err);
+        } else {
+          expect(cToken.pseudo).to.equal(user.pseudo);
+          token = cToken;
+          done();
+        }
+      });
+    });
+
+    it('Should create a new cast token with an anonymous user', function (done) {
+      common.castTokenModel.createNew(client, null, cast1, true, function (err, cToken) {
+        if (err) {
+          done(err);
+        } else {
+          expect(cToken.pseudo).to.not.be.undefined;
           done();
         }
       });
@@ -147,8 +248,42 @@ describe('Testing Spredcast models', function () {
         if (err) {
           done(err);
         } else {
-          expect(fCast).has.lengthOf(1);
+          expect(fCast).has.lengthOf(2);
           done();
+        }
+      });
+    });
+  });
+
+  describe('castTokenModel.getByToken()', function () {
+    it('Should find the created token', function (done) {
+      common.castTokenModel.getByToken(token.token, function (err, fToken) {
+        if (err) {
+          done(err);
+        } else {
+          expect(fToken._id.toString()).to.equal(token._id.toString());
+          expect(fToken.user.pseudo).to.equal(user.pseudo);
+          expect(fToken.cast.name).to.equal(cast1.name);
+          done();
+        }
+      });
+    });
+  });
+
+  describe('castTokenModel.revokeCast()', function () {
+    it('Should remove all token generate for cast1', function (done) {
+      common.castTokenModel.revokeCast(cast1._id, function (err) {
+        if (err) {
+          done(err);
+        } else {
+          common.castTokenModel.getByToken(token.token, function (err, fToken) {
+            if (err) {
+              done(err);
+            } else {
+              expect(fToken).to.be.null;
+              done();
+            }
+          });
         }
       });
     });